Reverted to edit edt-dd1f15b5-9554-45aa-91a8-a7ca6c42e3d2: "Implement book preview feature
This commit introduces the "Preview Book" feature, allowing users to view a digital representation of their storybook. The preview includes cover, chapters, multimedia, interactive navigation, zoom, pan, and bookmarking functionalities. It fetches the latest book data from the database and reflects real-time updates."
