+++ package.json
@@ -46,10 +46,8 @@
     "cmdk": "^1.0.0",
     "date-fns": "^3.6.0",
     "embla-carousel-react": "^8.3.0",
-    "html2canvas": "^1.4.1",
     "input-otp": "^1.2.4",
     "js-cookie": "^3.0.5",
-    "jspdf": "^3.0.0",
     "lucide-react": "^0.462.0",
     "next-themes": "^0.3.0",
     "plyr-react": "^5.3.0",
