+++ src/components/book/TableOfContents.tsx
@@ -10,15 +10,13 @@ interface TableOfContentsProps {
   currentPage: number;
   onSelectPage: (pageIndex: number) => void;
   bookmarks: number[];
-  storyPages: number[]; // Added this prop
 }
 
 const TableOfContents = ({ 
   stories, 
   currentPage, 
   onSelectPage, 
-  bookmarks,
-  storyPages 
+  bookmarks 
 }: TableOfContentsProps) => {
   return (
     <div className="space-y-4">
