+++ src/components/storybook/StoryBookCard.tsx
@@ -1,4 +1,3 @@
-
 import { Link } from "react-router-dom";
 import { formatDistanceToNow } from "date-fns";
 
@@ -17,7 +16,7 @@ export function StoryBookCard({ storybook }: StoryBookCardProps) {
   return (
     <Link
       to={`/storybooks/${storybook.id}`}
-      className="block border rounded-lg p-6 hover:shadow-lg transition-shadow bg-white"
+      className="block border rounded-lg p-6 hover:shadow-lg transition-shadow"
     >
       <h2 className="text-xl font-semibold mb-2">{storybook.title}</h2>
       {storybook.description && (
