+++ src/index.css
@@ -1,4 +1,3 @@
-
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
@@ -7,144 +6,4 @@
   .text-shadow {
     text-shadow: 0 2px 4px rgba(0, 0, 0, 0.5);
   }
-  
-  .animate-slide-in-right {
-    animation: slideInRight 0.3s ease-out forwards;
-  }
-  
-  .animate-fade {
-    animation: fade 0.3s ease-out forwards;
-  }
-  
-  .overflow-ellipsis {
-    text-overflow: ellipsis;
-    white-space: nowrap;
-    overflow: hidden;
-  }
-}
-
-@keyframes slideInRight {
-  from {
-    transform: translateX(-100%);
-    opacity: 0;
-  }
-  to {
-    transform: translateX(0);
-    opacity: 1;
-  }
-}
-
-@keyframes fade {
-  from {
-    opacity: 0;
-  }
-  to {
-    opacity: 1;
-  }
-}
-
-/* Card styles */
-.bg-card {
-  background-color: white !important;
-}
-
-/* Book preview styles */
-.book-page {
-  /* 5x8 inch book format with proper aspect ratio */
-  aspect-ratio: 5/8;
-  width: 100%;
-  height: 100%;
-  background-color: white;
-  overflow: hidden; /* Prevent scrolling */
-  position: relative;
-}
-
-.book-content {
-  /* Standard book content width with margins */
-  max-width: 90%;
-  height: 100%;
-  margin: 0 auto;
-  overflow: hidden; /* Ensure no scrolling */
-  position: relative;
-}
-
-.book-text {
-  /* 14pt font size - equivalent to roughly 18.5px */
-  font-size: 18.5px;
-  line-height: 1.5;
-  max-height: 100%;
-}
-
-/* Mobile-friendly adjustments */
-@media (max-width: 768px) {
-  .book-text {
-    font-size: 16px;
-    line-height: 1.4;
-  }
-  
-  .book-content {
-    max-width: 95%;
-  }
-}
-
-/* Ensure media content stays within page bounds */
-.book-page img, .book-page video {
-  max-width: 100%;
-  max-height: 25%; /* Limit height of media to avoid overflow */
-  object-fit: contain;
-}
-
-/* Page navigation indicators */
-.page-indicator {
-  position: absolute;
-  bottom: 1rem;
-  right: 1rem;
-  font-size: 0.75rem;
-  color: #666;
-}
-
-.page-continuation {
-  font-style: italic;
-  text-align: right;
-  margin-top: 1rem;
-  color: #666;
-  font-size: 0.875rem;
-}
-
-/* Book dimensions based on standard 5x8 inch format */
-.book-dimensions {
-  width: 480px; /* 5 inches at 96dpi */
-  height: 768px; /* 8 inches at 96dpi */
-}
-
-/* Mobile-friendly book dimensions */
-@media (max-width: 768px) {
-  .book-dimensions {
-    width: 100%;
-    height: auto;
-    max-height: 80vh;
-  }
-}
-
-/* Responsive padding and spacing */
-.responsive-container {
-  padding: 1rem;
-}
-
-@media (min-width: 640px) {
-  .responsive-container {
-    padding: 1.5rem;
-  }
-}
-
-@media (min-width: 768px) {
-  .responsive-container {
-    padding: 2rem;
-  }
-}
-
-/* Improve touch targets for mobile */
-.mobile-touch-target {
-  min-height: 44px;
-  min-width: 44px;
 }
